+++ SimulationModel.cpp
@@ -25,7 +25,7 @@ Full license text: https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html
 #include <cmath>
 
 size_t SimulationModel::GetMemSize() {
-    int modelSize = 0;
+    size_t modelSize = 0;
     for (auto &fac : facets)
         modelSize += fac->GetMemSize();
     for (auto &struc : structures)
