+++ SimulationModel.h
@@ -46,7 +46,7 @@ public:
     std::string name;
 
     size_t GetMemSize(){
-        int sum = sizeof(name);
+        size_t sum = sizeof(name);
         if (aabbTree) {
             sum += aabbTree->GetMemSize();
         }
@@ -145,7 +145,7 @@ public:
     virtual double GetOpacityAt(SimulationFacet *f, double time) const {return -1.0;};
     virtual double GetStickingAt(SimulationFacet *f, double time) const {return -1.0;};
 
-    int memSizeCache=0;
+    size_t memSizeCache=0;
 
     // Geometry Description
     std::vector<std::shared_ptr<SimulationFacet>> facets; // All facets of this sim model. Shared pointer as these facets are referred to in the ray tracing accel structure
